// .lovable/backup/stable-pre-refactoring/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5";
  };
  public: {
    Tables: {
// ...
      hashtag_follows: {
        Row: {
          created_at: string;
          hashtag_id: string;
          user_id: string;
        };
        Insert: {
          created_at?: string;
          hashtag_id: string;
          user_id: string;
        };
        Update: {
          created_at?: string;
          hashtag_id?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "hashtag_follows_hashtag_id_fkey";
            columns: ["hashtag_id"];
            isOneToOne: false;
            referencedRelation: "hashtags";
            referencedColumns: ["id"];
          },
        ];
      };
      hashtags: {
        Row: {
          created_at: string;
          id: string;
          label: string;
          last_used_at: string | null;
          posts_count: number;
          tag: string;
        };
        Insert: {
          created_at?: string;
          id?: string;
          label?: string;
          last_used_at?: string | null;
          posts_count?: number;
          tag: string;
        };
        Update: {
          created_at?: string;
          id?: string;
          label?: string;
          last_used_at?: string | null;
          posts_count?: number;
          tag?: string;
        };
        Relationships: [];
      };
// ...
      post_hashtags: {
        Row: {
          created_at: string;
          hashtag_id: string;
          post_id: string;
        };
        Insert: {
          created_at?: string;
          hashtag_id: string;
          post_id: string;
        };
        Update: {
          created_at?: string;
          hashtag_id?: string;
          post_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "post_hashtags_hashtag_id_fkey";
            columns: ["hashtag_id"];
            isOneToOne: false;
            referencedRelation: "hashtags";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "post_hashtags_post_id_fkey";
            columns: ["post_id"];
            isOneToOne: false;
            referencedRelation: "posts";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};